+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG.md
+## [2.16.0] - 2024-11-08
+Added:
+- Screen Sharing
+- File transpilation
+
+
 ## [2.15.0] - 2024-10-22
 Added:
 - Screen Sharing
@@ -10,6 +16,10 @@ Added:
 - updated typescript definition for AttributeDictionary, SegmentAttributeDictionary, ReferenceDictionary
 - typescript SYSTEM and ERROR types for AgentStateType
 - Agent and Contact events subscriptions now return subscription
+
+## [2.15.0] - 2024-08-16
+Added:
+- Global Resiliency implementation. See Documentation-GR.md for details.
 
 ## [2.14.5] - 2024-08-02
 Added: